+++ migrations/2020-03-25-135000_create_capabilities/up.sql
@@ -1,7 +1,7 @@
 -- Your SQL goes here
-CREATE TABLE roles (
+CREATE TABLE capabilities (
   id INT UNSIGNED NOT NULL AUTO_INCREMENT,
   name VARCHAR(150) NOT NULL,
   PRIMARY KEY (id),
-  CONSTRAINT unique_roles_name UNIQUE (name)
+  CONSTRAINT unique_capabilities_name UNIQUE (name)
 );