+++ src/database/models/mod.rs
@@ -1,2 +1,3 @@
 pub mod address;
+pub mod roles;
 pub mod user;